requirements.txt
atomicwrites >=1.1.5
attrs >= 18.1.0
coverage
jsonpatch >= 1.23
more-itertools	>= 4.2.0
pip	>= 10.0.1
pluggy	>=	0.6.0
psycopg2	>=	2.7.5
py	>=	1.5.4
pytest	>=	3.6.3
python-dateutil	>=	2.7.3
python-json-logger >=0.1.9
setuptools	>=	40.0.0
six	>=	1.11.0
# testing.postgres >= 1.3.0  # this needs to be commented out for deployment to AWS Lambda